Create MainPage.xaml.cs
This is the main code behind file in C# which consists of the code to import csv files. The CSV file selected using the dialog box is parsed and then assigned to the DataGrid using the UI virtualization technique in Silverlight.
The code for accessing the server side is also there in this file.